# FEXCore: Add non-atomic Memcpy and Memset IR fast paths