@bylaws 2024-03-18
Fixed the issue with backwards copies